return route details
make-routes now returns details for routes.
will parse the details (they are now just the raw jsons from the api)
so that they are usable for our purposes (having for each route, the route geometry and the route bus stops)
also. wondering if there is a more efficient way for/than throttling on each api call.