add points, return segs
addPoints method in turf-helpers exists and allows to generate segments in finer detail. This, again, is needed to find approximately close locations between route points and bus stops.
with this implemented, the segments code was extended to return an object for each route. these objects' values are lists with objects representing segments along routes.
: