fixed lat/lon + gen output
fixed my dumb lat/lon mixup.
We also now have the segments beinig made and written out to a file.
However, the way it is being done now (making thousands of points along the route then selecting points in that route closest to bus stops) is more than likely prohibitively inneficient. Not so much that file sizes will get too large (we can make things vector tiles), but building the routes will be slow, slow, slow. Looking at the turf documentation, one possible solution is to for each route using a mixture of the sector method and midpoints to location on a line between the closest two points to a route perpendicular to the bus route's location.