make addpoints include a loop
addpoints now is a loop of length 'loop', meaning it will add midpoints between two points in the route passed 'loop' number of times it is told to.
I explored other turf methods but they all seem to rely on the points that make up linestrings' geometry. This is not helpful for my purposes b/c the points along the routes I am trying to slice are...inprecise. I will loop back at my original while loop method for creating these routes and try to implement it. Need to wrap this up b/c more than likely this is work that only matters for display purposes... Given some of the other methods available with turf (most interesting of which is sector, I may only need the bus points to figure out on which segment busses are located.