working out new method for segments
working on a new method for generating segments.
since making a a large # of new points is inefficient, the new method tries the following. it first finds the closest point to bus stops on the for two segments in a stop, then using a gemetry made by two bearings that face stops' stop partner, selects points on the line.
This has its own problems, but was a good learning experience as it may be the answer with some more work, or stuff I learned to do (like use point on a line) might be able to make the segs we want.